Converted CSS to Sass
I decided to convert my CSS to Sass so that I could use this as an
opportunity to learn Sass. It's working most of the way, however the
page has a lot of styles missing, likely because I added a reset CSS
file, so this will be my next goal.